WIP - Alternative terminal point implementation for 'no orthogonal segments'
Only use the implementation involving the perimeter projection when there is no
waypoint. Otherwise, use the original first and last waypoints.
This avoids side effects when terminal waypoints are inside the shapes and - in
some cases - outside shapes.

Drawbacks: as the terminal waypoints can be inside or outsides shapes, the
rendering of such edge cases may be less pretty than before.