[TEST] Add visual tests for edges targeting collapsed elements (#1867)
Changes
  - add diagrams for expanded/collapsed pools and subprocesses
  - HTML page: manage collapse element query parameter
  - added tests: test the diagram rendering with and without collapsed elements

Note that collapsing elements is not officially supported and is managed with mxGraph custom code.
We demonstrate it in the examples, so the new tests ensure the sequence and message flows are rendered correctly. The new tests should avoid finding a problem in the examples repository after a new version is released.