[INFRA] Fix the npm version script to make it work with the GH workflow
Remove call to the cross-env command as it is not available at runtime in the release GitHub
workflow.
Dependencies are not available as we are not running `npm ci` prior running `npm version`.
We are now using the Linux syntax which is not an issue as we always run the npm script from the
GitHub workflow (never locally).