[DOC] Simplify section levels to manage both include and standalone pages
Previously, when writing in an included page, writer has to remind to set the levels as expected in the final document.
It meant to decrease all levels.
Now, the page can be written as it is a single document and the levels are decreased in the primary document to match
the value for inclusion.