[FIX] Display conditional marker from gateways
The style name was correctly set on cells in the mxGraph model, but we had not configured the style globally to display
the marker. It was correctly configured for the marker from activities.

A dedicated diagram has been added to ensure that the behaviour in now correct (e2e tests). It checks both horizontal
and vertical sequence flow markers.
The 'complex path' diagram has also been updated to check the marker rendering when the sequence flows are oblique.

Breaking changes
In `StyleDefault`, `SEQUENCE_FLOW_CONDITIONAL_FROM_ACTIVITY_MARKER_FILL_COLOR` has been renamed into
`SEQUENCE_FLOW_CONDITIONAL_MARKER_FILL_COLOR` as it applies to both sequence flows from activities and gateways.
Remember that using this constant to change default colors is experimental and subject to change without prior notice.