[FEAT] Add a new API to zoom the BPMN diagram (#1988)
The API lets do zoom in and zoom out from the center of the `bpmn-container`.

Introduce the `Navigation` class that exposes all diagram navigation features. It also includes the existing `fit`
function. The former one that is defined in `BpmnVisualization` class is deprecated to not break compatibility. It will
be removed in a future release.

IMPORTANT: the new Zoom API doesn't go through the throtlle/debounce mechanism that exists when zooming with the scroll wheel.

The demo and the diagram test page displays zoom buttons to demonstrate the usage of the new API.
In the demo page:
  - the zoom buttons replace the zoom debounce/throttle settings. They were only displayed for information and weren't
  useful for the end user.
  - some elements like labels could previously get focus. This is now fixed.

Refactor
  - In BpmnGraph, the `cumulativeZoomFactor` has been renamed into `currentZoomLevel` for clarification.
  Rationale: in mxGraph code, `factor` properties are not naming scale values, but a value used to multiply the scale.
  Here, we use something related to the scale, so we couldn't keep using the `factor` name.
  - In e2e tests, clarify the signature of the `pageTester.mouseZoom` utils method:
    - Use zoomType instead of `deltaX` values. Arbitrary `deltaX` values were used at several places. The value is too
    much tight to the mouse event implementation. The zoom type is now clearer and this reduces usage of duplicated
    constants.
    - Put the xTimes parameter at the end of the signature to allow default values. This also makes the signature
    consistent with the new `doZoomWithButton` function.