EXTRA rewrite wrong process test about participant without processRef
The test was wrong as it stated that a participant without processRef is never retrieved whereas this was because of
the lack of shape in the json. This is in fact what we call 'black box' pool and we already had test about this.
The test has been renamed and rewritten to show the behavior when the json contains both white and black box pools -
this was a missing use case.