[TEST] Ensure consistent page and directory names
We had different names that make it hard to see at a glance the purpose of the assets:
  - BPMN diagrams and snapshots directories didn't always have the same name
  - the name of the 'non-regression' page didn't state that it was intended to test the BPMN element rendering
specific config for bundle test

Changed elements
  - page
    - non-regression --> bpmn-rendering
  - bpmn diagrams
    - diagram --> fit
    - non-regression --> bpmn-rendering
  - snapshots
    - bpmn --> bpmn-rendering

Also simplify the utils used in e2e tests to select the page for the test. This avoids duplication and will simplify
the maintenance in the future.