[INFRA] Improve surge preview workflows (#2072)
Remove duplication by introducing a custom action. We now run a single job instead of 2 jobs with cryptic secret checks.
Delegate most of the work to the https://github.com/bonitasoft/actions `surge-preview-tools` action to better manage PR created from fork repositories or by Dependabot.
We shouldn't have error on PR close when it was created by dependabot: most of the time the preview doesn't exist. We previously had errors because we always tried to teardown the surge domain even when it didn't exist. This is now fixed.