[DOC] Improve the doc of API marked as experimental in the HTML docs (#2095)
Typedoc 0.23 changed the way it renders 'experimental' tags. 
Previously, we could have a message explaining why it was experimental. Typedoc did display the message close to the tag. 
Now, it put the tag close to the class/property name and the message is display anywhere in the description, which makes the doc hard to understand. Some API documentation were already updated as part of the Typedoc version dump, but some remains unchanged.