[TEST] Make performance tests work (#2092)
We haven't run the test for a long time, and they failed due to configuration issue.
The jest-extended configuration was missing. It is indirectly required as it is used in e2e test utilities that performance
tests rely on.

In addition, the following refactoring have been done
  - improve the suite organization: we are running the same test several times. It previously had the same name
  independently of the run. They now have dedicated name, so it is easier to follow test execution.
  - centralize where the path to the data file is computed. It is now possible to run the tests in IntelliJ/Webstorm

Also add a new GitHub workflow to run the performance tests. It will help detecting runtime errors earlier. It isn't intended to store the performance results as the performance of the runners are not guaranty over the time.