[FEAT] Allow to fill the boundaries of the Text Annotation with color
Previously, only a small area of the Text Annotation was filled (the area delimited by the open rectangle).
Now the whole boundaries are filled which makes the fill setting useful. The whole area under the text can be filled
for people that want to highlight it.

In the tests/examples, introduce a new 'Light Blue' theme that demonstrates how text annotations with fill color are now
rendered.
The BPMN diagram used in theme visual tests has been improved:
  - only includes a single black box pool (having 2 black box pools didn't test anything special)
  - the removed black box pool has been replaced by a participant without shape: this lets test the fill color of the
  elements when there is no lane with a dedicated fill color.
  - the new participant includes an expanded subprocess and a collapsed call activity which were not tested in the past