[INFRA] gha - Use surge-preview-tools action v1 (#2185)
In the internal `custom-surge-preview` action, use the v1 version of `surge-preview-tools`. This lets benefit from the latest compatible improvements without having to update when a new version is available.
The latest 1.3.1 version improves logs: no log warnings when the surge token is not set (for dependabot PR or PR created from fork repositories), log the version of the surge CLI (debug).