[FIX] Filter elements of an Expanded Call Activity (#2140)
The filtering code already managed this use case and unit tests existed. To work, it needs that the parent relationship between the Call Activity and its elements was set.
Previously, the parsing implementation didn't set the relationship correctly. Instead of referencing the call activity, it referenced something else. So the elements were dropped when doing the filtering.

Now, the parsing code links the BPMN elements of an expanded Call Activity to the Call Activity via parent relationship.
Then, when a pool containing a Call Activity is filtered at load time, the elements of the Call activity are filtered as well.

The parsing unit tests and the integration model tests have been improved to ensure we check the inner elements of the expanded Call Activity
We already had a visual tests about the filtering of call activity elements that shown the past issue. They now show it is fixed.