[FIX] Filter elements of a participant without a displayed Pool (#2139)
The BPMN diagram can contain participants without `BPMNShape`. In this case, the pool is not displayed, but the BPMN elements related to the participant are displayed.
Previously, it was not possible possible to filter such elements. A `No matching pools` error was thrown. This is now fixed.

In addition, to improve the functional coverage of the filter feature with new unit tests:
- add for not displayed pool containing BPMN elements (demonstrate that the bug is fixed)
- add for black box pool filtering.

Co-authored-by: Olan anesini <olan.anesini@bonitasoft.com>