[DOC] Do not add dev dependencies to release notes
Dependabot now creates Pull Requests that are not present in the release notes.
As there is only runtime dependencies in the release notes, the dependency updates list isn't collapsed to ensure all
changes are visible.
The `Dependencies update` section is now above the `Maintenance` section as dependencies update are more important for
users than what we did internally to improve the library.

Other release notes improvements:
  - exclude process-analytics-bot. The project bot is excluded from the $CONTRIBUTORS variable to avoid manual edits of
  the release notes.
  - update after label changes in the GitHub repository
    - infra:refactoring --> refactoring
    - infra:build --> chore
    - infra:repo: removed