[FEAT] Simplify the integration in TypeScript project
TS projects integrating the lib no longer need to declare 'typed-mxgraph' types. The TS compilation now works out of the
box.

Previously, the declaration files included in the npm package referenced 'typed-mxgraph' with:
    /// <reference types="typed-mxgraph" />

"typed-mxgraph" had to be accessible directly and so, "node_modules/@typed-mxgraph" had to be added to TS configuration
`typeRoots`. This was because the lib itself declared it in `typeRoots`.

The changes are tested with a dedicated TypeScript project that runs the TS compiler. It also tests the lowest TS
version that can work with the lib (currently 4.5).

The README has been updating accordingly. It contains a note about the configuration that was previously required. It
should help projects still using an old version.