[FEAT] Simplify the integration in TypeScript project (#2258)
TypeScript projects integrating the lib no longer need to declare
'typed-mxgraph' types. The TypeScript compilation now works out of the
box.
Projects that customize the mxGraph code don't need additional
configuration to access mxGraph types.

Previously, the declaration files included in the npm package referenced
'typed-mxgraph' with:
```
    /// <reference types="typed-mxgraph" />
```

"typed-mxgraph" had to be accessible directly and so,
`node_modules/@typed-mxgraph` had to be added to TypeScript
configuration `typeRoots`. This was because the lib itself declared it
in `typeRoots`.

The changes are tested with a dedicated TypeScript project that runs the
TypeScript compiler. It also tests the lowest TypeScript version that
can work with the lib (currently 4.5).

The README has been updating accordingly. It contains a note about the
configuration that was previously required. It should help projects
still using an old version.