[FEAT] Explicit error message when using too old TypeScript versions (#2272)
TypeScript projects integrating `bpmn-visualization` but using an
unsupported TypeScript version now get an explicit error message instead
of a list of type errors.

For instance, when using TypeScript Version 4.4.4
```
node_modules/bpmn-visualization/dist/not-supported-ts-versions.d.ts:1:74 - error TS1003: Identifier expected.

1 "Package 'bpmn-visualization' support only TS versions that are >=4.5.0".
```

Related changes
- The documentation provide general guidelines to contributors about the
supported TypeScript versions and how changed them.
- As types are now generated in several steps, Rollup is no more in
charge of generating the types. Rollup was previously generating the
files several times (each time it generated a bundle so 3x2=6 times), so
this was a lost of time.
- Rename the tsconfig.json file used to prepare the npm package. It now
has a clearer name.
- Do not generate package-lock.json in the typescript-support tests and
remove it from git. The version of bpmn-visualization should be updated
in this file after each release. We don't want to do it, so avoid to
generate the lock file