[INFRA] Run some GH workflows less often
Currently, we can run at most 20 jobs simultaneously and no more than 5 in parallel on macOS.
To reduce the load, reduce the amount of jobs that are run when pushing.

The main change is to stop running bundle tests on all OS with all browsers. This drops 11 jobs for each PR that runs
the 'build' workflow! The bundle tests are now run in the workflow that build and upload the npm package.

Other configurations of workflows have been updated to not run when some files not involved in the workflow run are
modified.