[DOC] Simplify the content of the NOTICE file
Remove extra TSDoc syntax, only keep plain text.
The NOTICE file has been excluded from the eslint processing: this file doesn't change, is related to the documentation,
and does not follow the JS/TS header file format. So there is no reason to lint it.

Restore the Copyright date to `2020`. This is the start date of the project.