[REFACTOR] Include a single file for types in the npm package (#2325)
The file is generated thanks to `api-extractor` which use the types
generated by `tsc` to build a single file including all types.
This is more consistent with the ESM bundle: we now have a single file
for both types and ESM in the npm package, and the internal layout of
the sources is not exposed externally.
As a bonus, it removes extra types that were erroneously available in
the npm package (JSON model types).

The bundle tests have been divided with a file that tests the display in
an HTML page using the IIFE bundle and a file that checks the files
available in the npm package.
For the later, the checks are stricter than before. They expect an exact
list of files instead of just checking that some files are present (it
missed extra files).

Note that the types now export an extra constant (`mxgraph`, holding the
mxGraph value objects) that is not exported in the ESM bundle. It could
lead to integration errors: import this constant in the TypeScript
source and get an error at runtime as this constant doesn't exist in the
JavaScript bundle.
It is not possible to remove it for now without breaking the single file
generation. This will be fixed later.