[FEAT] Downlevel the minimum supported TS version from 4.5 to 4.0 (#2326)
Explicitly support version greater or equal to 4.0.2 (first GA version
of the 4.0 line). The README now explains what was the version supported
for older releases of `bpmn-visualization`. It shows the progress and
will help people using old versions to not be misguided when looking the
documentation directly from the GitHub Web UI or the README of the
`master` branch.

In addition, update the types file included in the npm package and used
when the TypeScript version is not supported. It now contains a comment
explaining what this file does.