[INFRA] Use chrome/edge installed in GitHub runners (#2353)
Previously, we installed Chrome and Edge prior running the tests
whereas the browsers are already installed in the GitHub runners.
This increases build time and limit build reproducibility (the available
version for download can change without notice).

Introduce an action to manage browser download logic, in particular, to
decide whether to download the browser. This will help in the future if
we want to cache browsers downloaded by playwright itself.