[TEST] Bump Jest dependencies to use version 29 (#2352)
jest-playwright-preset@2.0.0 (latest available version for now) only
declares jest@28 as peerDependencies, not jest@29. So, activate the
`legacy-peer-deps` option in `.npmrc` to avoid `npm install` errors.

Also remove the `@types/jest` dependency. We declared it because we use
ts-jest and it is already declared by ts-jest. So, stop maintaining this
dependency declaration.