[REFACTOR] Clarify logic of the style configuration for `Associations`
In all `AssociationDirectionKind` styles, add configurations for end and start arrow, instead of removing it. This is
consistent with the rest of the configuration and make the configuration easier to follow.

Also introduce a constant to have same size by default for all end and start arrows.
This avoids the use of the magic number several times. Instead, the constant with a convenient name makes the
configuration clearer.