[REFACTOR] Improve the default style configuration (#2462)
Shape: move the "arc size" configuration to the default vertex style:
  - This avoids duplication in specific styles
- The configuration is now inherited from the default style, so it is
easier to change it. Just change it on the default, and it will be
applied automatically to all vertices

Default edge rounded settings: used boolean instead of number for
consistency. Boolean are used everywhere else.

Shared `configureCommonDefaultStyle` function:
- Call it in first. This lets override any values it defines in specific
default edge and vertex configuration.
- Move it out of the class as it doesn't depend on any values and
properties of this class