refactor: introduce `batchUpdate` for model transaction
This new method in `BpmnGraph` avoids using the try/catch idiom which is replaced by a simple method call.
It is inspired from the related `maxGraph` function.

Also add the model clear within the transaction in `BpmnRenderer`. This increase the overall speed because the model
changes are all "committed" at the same time. We can then expect better overall rendering performance.

In addition, always use the `panningHandler` const in `GraphConfigurator` for consistency with the rest of the code and
to avoid any future side effects.