refactor: remove StyleUtils (#2550)
Replace `StyleUtils` methods by direct calls: the methods were used only
once, so this was an extra indirection (the inline code is explicit
enough).

Extract only the `getBpmnIsInstantiating` function because it is used
twice and make it internal.

It also changes the way we store `bpmn.isInitiating` in the mxGraph
style. Previously, we stored the value from the BPMN Semantic as a
string which required a subsequent parsing to know whether it was an
"instantiate" value. This was also not consistent with the property name
which suggests to store a boolean value.
We now store a boolean value which can be directly used in the Shapes
implementation.

--------------------------------------------------------------------------------
BREAKING CHANGES: `StyleUtils` may have been used in rare cases to
redefine the way the shapes are rendered. It wasn't used in the examples
of `bpmn-visualization`.
`StyleUtils` was mark as experimental and as subject to change as part
of BPMN Theme refactoring. So users already knew that a deletion could
take place.