test: improve the layout of the integration test BPMN Diagram (#2554)
The message flows did not declare waypoints, so they passed through many
shapes. The diagram was not easy to read. These flows now bypass all
shapes.
In addition, a message flow wasn't displayed because of a bad id set in
its targetRef. Also reverse the source and target to have a better semantic.