refactor: rework label position setting in StyleComputer (#2551)
Previously, we set invalid values for STYLE_LABEL_POSITION and
STYLE_VERTICAL_LABEL_POSITION but the rendering was correct.
For STYLE_VERTICAL_LABEL_POSITION, the position doesn't change if is not
'bottom' or 'top'. That's why passing an invalid value (i.e. a string
different from 'middle'). We now pass 'middle' which as the same
rendering effect but is valid according to the mxGraph documentation.
For STYLE_LABEL_POSITION, we rely on an undocumented mxGraph feature.
The value is supposed to be set to one of 'center', 'left' or 'right'.
But, when it is set to something else, the rendering is done like we
want (no change in the label position). So, to clarify the code, we use
'none' instead of 'top'.

Also clarify the comment about how we set the STYLE_HORIZONTAL mxGraph
style because there is not the same meaning in BPMN and in mxGraph.