ci(e2e-tests): fix false positive tests on macOS with Firefox (#2562)
The tests were previously blocked if a fetch diagram error occurred. A
`window alert` appeared, which prevented Playwright from interacting
with the browser afterwards (for retries and subsequent test runs).

As part of the fix:
- add logs to improve troubleshooting
- do not display `window alert` in test pages when an error occurs on
diagram fetch ("diagram-navigation.html" and "overlays.html"). This
blocked the browser and all following tests finished in error
(Playwright is unable to interact with the page).
- also avoid `window alert` on BPMN load error in the test pages.
- change the error management to only log errors by default. If a new
page is created, it won't use `window alert` by mistake.

Fetch errors can occur but this is not an issue as we configure jest
retries. Generally, the second time the test runs, there is no fetch
error.
About the `window alert`:
- the test and the demo pages share some code for error management
- the demo pages currently display errors in the `window alert` (this
will be improve in the future, by displaying the errors in a modal for
instance)
- the test pages should have only log errors or display the errors in a
dedicated element like this is done in the "bpmn-rendering.html" page.
The error management wasn't properly configured, so it used the default
which displayed errors in a `window alert`.