feat: introduce the `Update Style` API (#2556)
This is the first implementation which lets update the stroke color of
an Edge or a Shape.

The way the "rounded" style is configured for activities has changed. It
is no more defined in the mxGraph shape implementation but in the
mxStylesheet. Otherwise, the rounded borders were removed when updating
the style with the new API.

The JSDoc links the "Style Update" and the "CSS Class" APIs to remind
users that they both provide complementary ways to style BPMN elements.

Its secondary objective is to set up the infrastructure: API methods and
types, the integration tests and the e2e tests.
Some generalization will be done later when we start supporting style
update for more and more properties.
The integration tests contain duplication, especially between the shapes
and edges tests. This is done on purpose. The tests will be easier to
refactor when more style properties are supported (there is a better
chance to see a better abstraction emerge).

In addition, some utility code and documentation (closed to what was
changed here) were updated.
- logs in the demo/test code (for BPMN diagram load)
- JSDoc of the model check utility used in integration tests