test: simplify utility code in integration tests (#2572)
Clarify the name of the properties related to the style. They now
correspond exactly to their related counterpart in the mxGraph objects.
Rename utility functions to match the new name of the properties. The
tests of the "Update Style" API can now fully use the utility functions
as they also manage the style of the view state of the cell. So, this
reduces the duplication.