feat: reset style per property with the "Update Style" API (#2585)
Allow to reset the property individually. Some font properties cannot be
reset by this method, so the JSDoc explain why.

In addition:
- update the JSDOC of the updateStyle API: the method was introduced in
v0.33.0 not v0.32.0 (which is already released)
- for `color` properties, remove some special `mxGraph` keywords
(inherit, swimlane) because we don't test them. We also don't know
exactly how they behave with our library at the moment. It is still
possible to use them because all strings are allowed. But we don't put
them forward anymore.