feat: add support for 'BPMN in Color'
Support the 'BPMN in Color' and use bpmn.io colors as fallback.

This introduces internal management of BPMN extensions for BPMNShape, BPMNEdge and BPMNLabel.

The parsing tests are using diagrams taken from the 'BPMN in Color' specification sample and from the original issue
requesting support for modeling colors (#2588) as an example of bpmn.io colors.