docs: improve the README (examples + firefox logo) (#2670)
Examples: put the variable `bpmnContent` after the initialization of the
library. It was not necessary to put it so
prominently. Its new location allows to keep the declaration as close as
possible to its use in order to simplify the
reading.

Firefox logo: the URL previously used is no more reachable. Use a new
logo taken from the official mozilla.design
website.
We can expect more stability from this source which is the official
branding site for mozilla products.