refactor: transform overlay types from interface to type (#2675)
Previously, these types were defined as `interface`. The changes
suggests moving them to `type` aliases, aligning with TypeScript's
philosophy. This change does not affect the API contract and is aimed at
providing a better match to the intended usage of the structures, and a
consistent declaration with the `StyleUpdate` type.