test(e2e): only run the "Update Style" API when query parameters are set
Previously, the page always created an `StyleUpdate` object even when no related query parameters was set.
So the "Update Style" API was always called.
The object is now only created when query parameters are present. It lets test the rendering when the API is not
called (which is the general use case).