feat(parser): let configure XML attribute processing
Remove the 'entities' dependency.
By default, there is no need to be able to decode all entities as most of the time, only 'basic' entities are generally
used in BPMN sources. Instead, `bpmn-visualization` now provides a small set of entities transform.

If there is a need to transform more entities, a new Parser option is now available to let application perform more
processing.