feat(parser): generate more explicit warning logs (#2677)
In case of a parsing warning (unknown element for example), a warning
log is emitted in the browser console. Previously, it was hard to
understand that `bpmn-visualization` generated this log because there
was no indication about its origin.
Now, all logs are prefixed, so it is easier to see they are generated by
`bpmn-visualization`.