chore(deps): use the `lodash-es` dependency (#2678)
Previously, there were 2 CommonJS dependencies: `lodash-throttle` and
`lodash-debounce`.
It is now replaced by a single ESM dependency: `lodash-es`.

The change reduces the size of the minified IIFE bundle from 1.2kb.
With the CommonJS lodash dependencies, as the `throttle` implementation
calls `debounce`, `debounce` is present in `lodash-throttle`. In this
case, rollup (that generates the bundle) is unable to tree shaked the
duplication.
In the resulting bpmn-viusalization IIFE bundle, there are 2 `debounce`
implementation: debounce (from `lodash-debounce`)and debounce$1 (from
`lodash-throttle`).
Notice that the previous CommonJS dependencies only includes the
necessary code to run. The new `lodash-es` dependency includes the whole
lodash code which increases the size at installation on the developer machine.

The usage of a ESM dependency also provides a better interoperability
for bundlers. For example, Angular 15 warns usage of CommonJS
dependencies. In the past, applications integrating `bpmn-visualization`
had to add configuration to remove the warning emitted because of lodash
dependencies.
This also reduces the number of dependencies from 4 (2 libs and their
related types) to 2.