fix: keep CSS classes when calling the "Update Style" API (#2709)
Previously, calling the "Update Style" API after adding the classes with
the CSS API removed the CSS classes from a given BPMN element.
This is now fixed by setting the classes in the model instead of
updating directly the state. The state is refreshed when calling the
"Update Style" API and the manually added classes were removed.

Basic integration tests have been added to check that the CSS classes
are present in the style property of the mxGraph model. They only check
a few use cases to help detecting regression They complement the
existing tests checking that the classes are correctly set in the SVG
nodes. The existing tests cover all use cases of the CSS API.