feat: add support for 'BPMN in Color' (#2614)
Support the 'BPMN in Color' and use bpmn.io colors as fallback.
The BPMN parser always extract the colors, but by default, they are
ignored by the Renderer. It is possible to enable the support in the
Renderer by setting an options at the library initialization.

This introduces internal management of BPMN extensions for BPMNShape,
BPMNEdge and BPMNLabel.

The parsing tests are using diagrams taken from the 'BPMN in Color'
specification sample (this ensure that the implementation is compliant
with the specification) and from the original issue requesting support
for modeling colors as an example of bpmn.io colors.