feat: add support for 'transaction' subprocess (#2719)
This is a generalization of the management of regular sub-processes. In
the semantic of the BPMN source, transaction sub-processes are defined
by the "transaction" node whereas regular sub-processes are defined by
the "subProcess" node. The code and the tests have been updated to
handle both sub-process flavors. This will make the support of "adhoc"
sub-processes easier in the future (they are defined by the
"adHocSubProcess" node in the semantic of the BPMN source).

The changes also include the final rendering of transaction sub-process
BPMN elements.