test: simplify integration tests
General changes
  - Remove async directive when not needed
  - Fix the message produced by the "toBeEndEvent" function is case of failure: it previously mentioned "toBeStartEvent"

Improve typing guidance
  - Introduce type for horizontal and vertical align in ExpectedModelElement
  - Use a more precise type for ExpectedCallActivityModelElement.globalTaskKind

Use or introduce expect function targeting dedicated BPMN kind
  - this makes checks more explicit and simplify the configuration
  - introduce functions for inclusive and parallel gateway, group and text annotation
  - remove the too general "expectedShape" function

Introduce shared function for expect pool and lane and correctly expect the horizontal mxGraph style value.
It is for "swimlane" label only and is the opposite of the horizontal value in the BPMN Shape.
When the shape is horizontal (this is the regular way to model the diagram from left to right), the label of the
container is vertical.
When the shape is vertical (modeling from top to bottom), the label is horizontal.
This is correctly computed in the StyleComputer but the logic was inverted in the expectation of the integration tests.
This is now fixed. In addition, rename the property storing the expected value in "ExpectedShapeModelElement" to
clarify its purpose and avoid confusion with the BPMN Shape value.