feat: detect Ad-Hoc Sub-Processes (#2723)
The ad-hoc marker is not rendered for now but it is parsed, stored in
the internal model and used in the mxGraph cell style. The rendering
will be done later.

Also change the way markers are tested in integration tests.
Previously, markers were only tested with the raw mxGraph style defined
as a string. But the order of markers in this property is not
guaranteed, so it was difficult to correctly assume the value.
Instead, the utility code now only tests that the style property is set
in the case where markers are expected. In addition, the markers are
converted into an array whose contents are checked regardless of its
order. This is the same mechanism used in unit tests.