refactor!: make `StyleConfigurator` internal
`StyleConfigurator` was previously exported for customization of the BPMN theme. These customizations were closely
linked to the StyleConfigurator implementation, so they were fragile and could break quite often.
We now have simpler ways of configuring the theme, even if the final implementation is not available. So, we can make
`StyleConfigurator` internal.