refactor!: transform StyleDefault into an object (#2728)
`StyleDefault` was previously an enum but there is no need to enumerate
the values. In addition, the values store both number and string which
is generally considered confusing and a bad practice. Switching to a
regular object make things more explicit and fix several SonarCloud code
smells related to the enum usage (mix value types and non literal
values).