refactor: directly use mxGraph value objects (#2745)
The mxGraph initializer module now exports main mxGraph objects that are
used several times in the library. This make the code easier to read and
reduce the size of minified bundles.

So, when used, there is no more need to prefix the call of mxGraph
object by mxgraph. For instance, instead of writing
`mxgraph.mxConstants.XXX`, it is possible to write `mxConstants.XXX`.
Notice that in files where both values (from mxGraph) and types (from
typed-mxgraph) are imported, the types are aliased to avoid name clash.
The convention taken is to postfix the type by `Type` to make it
explicit.