fix: apply style update to the message flow icon (#2744)
The message flow icon is stored in a dedicated cell, but the style of
this cell had not been updated. The style update now always applies,
individual style properties can be set to default values, and the entire
style can also be reset.

The integration test utility code has been improved and generalized to
better check the message flow icon cell in the mxGraph model. In
particular, this verifies that the style is now correctly updated.